requirements.txt
google-generativeai
flask
flask-script
gunicorn
python-dotenv

# CrewAI and Agentic AI dependencies
crewai>=0.28.0
crewai-tools>=0.4.0
langchain>=0.1.0
langchain-google-genai>=1.0.0
pydantic>=2.0.0

# Optional tools
pytest>=7.0.0
